requirements.txt
requests==2.23.0
psycopg2-binary==2.9.3
docker==5.0.3 # или подходящая версия для локального сетапа докера